+++ tcp/factory.go
@@ -9,8 +9,9 @@ import (
 )
 
 const (
-	readerDS = "ReaderDataStore"
-	writerDS = "WriterDataStore"
+	readerPlugin    = "ReaderPlugin"
+	writerPlugin    = "WriterPlugin"
+	schedulerPlugin = "SchedulerPlugin"
 )
 
 //
@@ -20,8 +21,8 @@ const (
 // factories factory container
 var factories = make(map[string]interface{})
 
-// createDS real factory method
-func createDS(conf map[string]string, key string) (interface{}, error) {
+// createPlugin real factory method
+func createPlugin(conf map[string]string, key string) (interface{}, error) {
 	defaultKey := ""
 	if got, ok := conf[key]; ok {
 		defaultKey = got
@@ -33,14 +34,32 @@ func createDS(conf map[string]string, key string) (interface{}, error) {
 		for k := range factories {
 			availableDatastores = append(availableDatastores, k)
 		}
-		return nil, ErrInvalidDataStoreName
+		return nil, ErrInvalidPluginName
 	}
 	return engineFactory, nil
 }
 
+// createScheduler real factory method
+func createScheduler(conf map[string]string) (cron.Scheduler, error) {
+	ef, _ := createPlugin(conf, schedulerPlugin)
+
+	if ef != nil {
+		fmt.Println(reflect.TypeOf(ef)) // debug
+		if fn, ok := ef.(func(map[string]string) (interface{}, error)); ok {
+			if ds, _ := fn(conf); ds != nil { // casting
+				return ds.(cron.Scheduler), nil
+			}
+		}
+		err := ErrCasting
+		log.WithFields(log.Fields{"err": err}).Error("Create scheduler")
+		return nil, err
+	}
+	return nil, ErrInvalidPluginName
+}
+
 // createWriterDS real factory method
 func createWriterDS(conf map[string]string) (psmb.IWriterTaskDataStore, error) {
-	ef, _ := createDS(conf, writerDS)
+	ef, _ := createPlugin(conf, writerPlugin)
 
 	if ef != nil {
 		fmt.Println(reflect.TypeOf(ef)) // debug
@@ -53,12 +72,12 @@ func createWriterDS(conf map[string]string) (psmb.IWriterTaskDataStore, error) {
 		log.WithFields(log.Fields{"err": err}).Error("Create writer data store")
 		return nil, err
 	}
-	return nil, ErrInvalidDataStoreName
+	return nil, ErrInvalidPluginName
 }
 
 // createReaderDS real factory method
 func createReaderDS(conf map[string]string) (psmb.IReaderTaskDataStore, error) {
-	ef, _ := createDS(conf, readerDS)
+	ef, _ := createPlugin(conf, readerPlugin)
 
 	if ef != nil {
 		fmt.Println(reflect.TypeOf(ef)) // debug
@@ -71,18 +90,18 @@ func createReaderDS(conf map[string]string) (psmb.IReaderTaskDataStore, error) {
 		log.WithFields(log.Fields{"err": err}).Error("Create reader data store")
 		return nil, err
 	}
-	return nil, ErrInvalidDataStoreName
+	return nil, ErrInvalidPluginName
 }
 
 // Register register factory methods
 func Register(name string, factory interface{}) {
 	if factory == nil {
-		err := ErrDataStoreNotExist
+		err := ErrPluginNotExist
 		log.WithFields(log.Fields{"Name": name}).Error(err.Error())
 	}
 	_, registered := factories[name]
 	if registered {
-		err := ErrDataStoreExist
+		err := ErrPluginExist
 		log.WithFields(log.Fields{"Name": name}).Error(err.Error())
 	}
 	factories[name] = factory
@@ -90,10 +109,15 @@ func Register(name string, factory interface{}) {
 
 // WriterDataStoreCreator factory method to create writer task data store
 func WriterDataStoreCreator(driver string) (psmb.IWriterTaskDataStore, error) {
-	return createWriterDS(map[string]string{writerDS: driver})
+	return createWriterDS(map[string]string{writerPlugin: driver})
 }
 
 // ReaderDataStoreCreator factory method to create reader task data store
 func ReaderDataStoreCreator(driver string) (psmb.IReaderTaskDataStore, error) {
-	return createReaderDS(map[string]string{readerDS: driver})
+	return createReaderDS(map[string]string{readerPlugin: driver})
+}
+
+// SchedulerCreator factory method to create scheduler
+func SchedulerCreator(driver string) (cron.Scheduler, error) {
+	return createScheduler(map[string]string{schedulerPlugin: driver})
 }